requirements.txt
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
fonttools==4.61.0
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
wcwidth==0.2.14
webencodings==0.5.1
wheel==0.45.1
yarg==0.1.9
